+++ tests/test_forStr.nim
@@ -86,5 +86,6 @@ suite "For string":
     ]
   test "strip":
     check "ááźtest heállÊo".strip(RunesWithAccent) == "test hello"
-    check "### `func almoçarComÁgua(comida, litros: string): string`".
-      strip(chars = NonExtendedAlphanumeric) == "funcalmoçarComÁguacomidalitrosstringstring"
+    check "`func almoçarComÁgua(comida, litros: string): string`".
+      strip(chars = NonExtendedAlphanumeric, @[' ']) ==
+        "func almoçarComÁguacomida litros string string"